+++ packages/core/src/typography/Typography.tsx
@@ -1,6 +1,6 @@
 import { m } from 'framer-motion'
-import { X, Y, zeroPosition } from '../general'
-import { getClassName } from '../themes'
+import { zeroPosition } from '../general'
+import { getClassName, ssrCompatible } from '../themes'
 import { TypographyProps } from './types'
 
 export const Typography = ({
@@ -14,14 +14,15 @@ export const Typography = ({
 }: TypographyProps) => {
     if (!children) return null
     const compositeClassName = getClassName(variant, className)
-    const config = { x: position[X], y: position[Y], rotate: angle, originX: '0px', originY: '0px' }
+    const [x, y] = position
+    const config = { x, y, rotate: angle, originX: '0px', originY: '0px' }
     const role = setRole && variant !== 'default' ? variant : undefined
     return (
         <m.text
             role={role}
             initial={config}
             animate={config}
-            style={style}
+            style={ssrCompatible(style, config)}
             className={compositeClassName}
         >
             {children}